Checking for non-zero array size before showing the results
Reference Issue : #854
For the case of search results with a single result item, metrics service treats it as a dataset landing page. This issue should be solved withing Metrics Service.

However, the UI should have a check if the response returned has contents in the array objects or not while displaying metrics. This commit takes care of that..