Inline citations and bibliography in markdown almost fully working.
The caveats are as follows: If a citation key isn't found in the bibtex, nothing is printed (the inline citation is just gone). Seems like instead I should return the citation key that was given. The other thing is that I'm not currently linking the inline citations down to the item in the bibliography. I'll try to fix those problems tomorrow. #870