Markdown citations working except of linking.
Inline citations are redered according to the csl style when found in bibtex and just rendered as the ciatation-label when not found. I would still like to figure out how to link the inline cite to the corresponding bib entry but I think that modifying the csl template may be the only way to make that work. I'll have to dig into it a bit more next week. #870