Inline citations linked to bibliography via ugly hack.
I had to hack the CSL template (embedded in showdown-citation.js) and then do a bunch of weird stuff with regular expressions. I couldn't embed the link tags directly in the CSL because, apparently, the processor strips it out. I'm not sure we'll really want to keep this code, because it won't work with any unmodified CSL templates. #870