Add CSS for new table-based attributeList layout
The change from a list to a table was to support having the badge next to each attribute name indicating it has a least one annotation and to have that badge stand out to the side of the name. I think this makes it a lot easier to spot the annotated attributes.