Changed the ordering of subview rendering so that the subview elements get inserted before render. This fixes a bug in the Google Maps rendering that requires the map canvas element to be in the DOM.
Added a template for the Metrics editor section

Separated all the section rendering into separate functions

Default to switching to the first tab listed in the editor instead of 
the Data tab

Save jQuery selectors to container elements on the view itself rather 
than inline in the code. This can make theming in MetacatUI easier.

Create a new Search model when the CollectionModel is initialized (in 
the defaults() hash), so that new Projects have one.

Create a new Filters collection when a Search model is initialized (in 
the defaults() hash), so that new Search models have one.

Related to #990