Display the Filters from the collection definition when editing a collection.
Hide them when viewing the collection in the ProjectDataView.

When a DateFilter or ToggleFilter is saved to the collection, parse and 
serialize it with special rules.

Fix some other Filter display issues

Created a `definitionFilters` attribute on CollectionModels for storing 
Filters that have already been serialized to the collection definition.

Related to #991, #993