Update creator and publisher properties in JSON-LD routine
Google now complains about untyped "creator"s and "publisher"s when I run our JSON-LD thorugh their Structured Data Testing Tool. This updates the implementation to assume all creators are of type Person and all publishers are of type Organization and type them as such. We have to make these hard assumptions because we're generating the JSON-LD entirely from the Solr index and we don't store enough information in Solr to know exactly which type (Person or Organization) each creator or publisher is. I think these are the most sensible defaults.

Closes #1074