Fix bug where EMLText was adding space to start of markdown content
The EMLText220 model was adding a space at start of all markdown content during parse. This meant each time a markdown page was edited in the portal editor, another space was added to the start of the content. Now the EMLText model only adds a space as a separator when concatenating multiple markdown elements.