Add LocalForage for draft saving and a Drafts view
This commit adds in basic support for saving metadata drafts using the LocalForage library.

Saves of the draft are triggered when either the EditorView or EntityView receive the "change" event
which triggers either a new draft to be created or an existing draft to be updated.

Drafts are viewable at an un-linked /drafts route which was designed so support staff could direct a
user with a failed submission to this page while communicating with them over email. Once on this page,
a user can see their drafts and copy the EML XML for the draft to share with the support staff they're
communicating with.

A big todo here I'd still like to do is to add support for removing drafts once an object with the same PID
is registered with the Member Node. I tried a few things but haven't had luck just yet.

Ref #216, #1242