Add support for replacing an item in the editor
Adds a Replace option to the data item dropdown which replaces an existing item
with one chosen with a file picker. The old DataONEObject instance is modified
instead of creating an entirely new object which results in:

  - An MNStorage.update() call being made to obsolete the existing item
  - The entity information in the package's EML to be retained except for the
entityName which is updated to match the new file's basename.

Also, you'll see two changes in DataONEObject which I had to make to get the REST
calls to fire correctly:

In the `url` method, first I found a bad call to this.hasContentChanges which
should've been called using .get. Second, the PID of newPid in the REST call to
MNStorage.update() was being used when it should be the old PID, as per the
API.

Ref #653