Add simple support for deep link to a /view page
The way to do this has changed a bit since #83 was created so this is a little more straightforward. First, we've changed MetacatUI's default route scheme to no longer use fragment identifiers so the URL fragment is now free to use and (2) we've established a canonical URI pattern for DataONE Objects of https://dataone.org/datasets/{dataset}#{sub-resource} for identifying sub-resources within metadata records. Now we can provide functionality where we try to scroll to the portion of a page describing a given identifier by checking for and retrieving the value of the `hash` property on `window.location`, heuristically attempting to find a portion of the page describing that identifier via a `data-id` attribute, and scrolling to it.

This isn't perfect because we need to scroll after MetacatUI successfully loads the View Service response and the view service response may or may not put a `data-id` attribute on various elements correspond to EML elements with `id` attributes but it works if we follow the pattern. This represents a pretty tight binding between the two systems but it does give us a way to scroll to _any_ identified element in an EML record (one with an `id` attribute) so long as the View Service injects a sanitized version of the ID.

I also added a small tweak to the "More info" link to make it set the `id` of the package item / metadata entity to the URL fragment to encourage users to discover this functionality. I think we could expose this in a more first-class way later if we like.

Ref #83