Prevent multiple-saves of data objects
I think the old logic here was based on the assumption of eager saving when
items are added to the editor. Now that replacing files delays the save until
the entire editor is saved, data models need to behave much more like metadata
models. So data models should only be saved when they have content changes and
are not in progress, complete, or in error. Note that I could make this
conditional simpler code-wise but I'm going to leave it as it is because it's
very clear.