Prevent race condition in DataPackage.save()
I found a race condition that was causing double-saves of package members in DataPackage.save().

The old logic was such that a race condition was possible. DataPackage.save() gets called repeatedly during the package saving process (save members, then save ORE) and decides which member objects to save every time it's called. The old logic for deciding whether a data item neededs saving was to check its hasContentChanges attribute. This attribute is set to true when we replace an item and set to false when the new item has been saved to the server. However, because of callback timings, DataPackage.save() often gets called a second time before the initial save of our new object succeeds, setting hasContentChanges to false.

This change simply adds the same logic we use to prevent double-saving of metadata objects to data objects. The conditional block is a bit messy but it's super explicit now.

Closes #1355