Disable replacing a data item when we don't have permission
Fixes #1393

This fixes a bug where a user could initiate a replace operation without permission to obsolete the item which causes downstream problems for the editor. Now, we assume we don't have permission and disable the replace functionality until we find out that we do have changePermission access. The DataItemView for each item is re-rendered a few times while the editor loads so the work is done here in the render method rathern than init.