Fix bug when changing name of replaced items in the editor
Fixes #1396

The no-op call to this.model.set("hasContentChanges"); looks like an accident for two reasons: (1) hasContentChanges is used to indicate when we need to update the object, not its sysmeta and (2) AFAICT a call to Backbone.Model.set without a value is a no op.

This feels a bit unergonomic but I couldn't come up with a cleaner way to have replacing files and renaming work together. Replacing an item sets the sysMetaUploadStatus to "c" to prevent the editor from trying to update sysmeta after updating the object so this change only re-sets sysMetaUploadStatus when we need just a sysmeta update, not an object update.