Tweak editor's replace feature to update EML ids
Closes #1397

Now we update the id attribute on the matching EML Entity for the item we're replacing when we do the replace. This keeps everything in sync so PROV and attribute editing doesn't break sometimes.

This adds a new convention around DataONEObject.updateID() so we can know before we save our replaced item what the PID will be and then update the EML entity to match it. updateID can now be called ahead of time (outside of a save) and, when it's called again normally during save, it won't update IDs again if it's already been called. Whether updateID has been already called depends on the value of the oldPid attribute which is set when updateID gets called. After we save objects now, we ensure we unset oldPid so a replaced item can be replaced again.