Reset oldPid attribute on EML and PackageModel ons ave
This fixes a bug described in https://github.com/NCEAS/metacatui/issues/1397#issuecomment-638492097. I failed to propagate my changes to updateID to these two models as well. This bug only occured if you submitted a package and didn't leave (route away) from the editor before submitting again. We weren't resetting state on the EML model after the successful save so the update call was using the old PID. This wasn't specifically to do with replacing but this branch did introduce the bug.