Fix an issue with the Query Builder and new isPartOf filters
In new portals, changes to the automatically created isPartOf filter
were not triggering change/update events. This meant that deleting
this filter had no impact on the search results, and the rule numbering
was not updated. The listener that updates the Search Model when the
Definition Model changes is now set in the Collection model, and the
isPartOf filter is added using an object of attributes instead of a
Filter model.

Relates to #1175, #1510