Patch funding validation to also check for award
Closes #1533

The change here checks for the presence of the award element by querying the project's DOM instead of just checking the attribute on the Backbone model like we do for the funding element because EMLProject doesn't currently parse the award element. Once that lands, we can simplify this validation code.