Change how Copy Citation control works
Closes #1583

Before, we were setting the text to be copied for the citation in a dataset attribute which breaks when the text has one or more double quotes in it, truncating the copied text at the first double quote.

I thought we could just encode/decode the string but I found an easier way to fix this: The Clipboard.js library lets you specify the text to be copied by delegating to the text() of another element on the page. The citation string is already contained on the page and we even save a reference to it in the `citationContainer` variable which is set view-wide. So we pass the selector in instead of the literal value and copy now works on citations with double quotes in them.