Increase width of AccessPolicyView modal
It now fills up to 80% of the window but shrinks when the window is narrower. This makes the individual access rules render on a single line and avoids weird wrapping issues.

I'll note that the CSS selector for this rule was .access-policy-view-container.modal before this change which didn't readily make sense to me since I didn't see where we set the modal class on the container. Everywhere I could find, we set the modal class on the view itself, which is a direct child of the container.