Fix a bug in creating object access policies
Fixes an edge case where the editor was creating a new accessPolicy for a pre-existing object that didn't have one set. This isn't desired because Objects don't have to have accessPolicy and the editor needs to respect this. Before this fix, the editor would try to apply any default access policies to existing objects it didn't have changePermission permisson on.