Apply access policy changes to resource maps too
I'm kind of inventing a system here but I expect we'll want to change it in future iterations. Rather than refactor the AccessPolicyView to support editing access policies for more than one object or buildig a new View that can do this, I just built a toggle on the AccessPolicyView to broadcast any changes broadcasting the change and handing off control to the DataPackage class to  handle saving and error handling.