Refactor MeasurementTypeView to support notfound/notspecific
This provides a way for the user to indicate they either couldn't find a term or the term they found wasn't specific enough to match. We use configurable sentinel values to indicate either and provide UI for each.

An alternative approach I considered but didn't get very far on was to use `additionalMetadata` to track the sentinel values. I chose the approach here mostly out of ease because hooking up additionalMetadata would require building out a lot of things don't currently have.