Fix bug in ncbo-tree widget rendering of a selection
This wasn't a bug until we added tooltip support to the tree elements. Adding tooltips involves adding the "tooltip-this" class and the old code only supported having a single class set on each element.

This was also the cause of some missing tooltips I hadn't quite nailed down so this is a twofer.