Tweak fallback description text, removing <a> tag
It looks like Dataset Search parses from HTML embedded into the JSON-LD it processes but I couldn't find an example of them processing an <a> tag. I did find an example of the automatically linkifying a bare URL in text so I'm tweaking the routine here to do that instead of wrapping the link in an <a> embedded in a string.