Forward /view/$data to /view/$metadata#data
Ref #1408

This changes the way getModel() works from recursively calling itself when the view is created for a non-metadata PID. Previously, getModel would be recursively called, calling itself again once it found the matching metadata PID for the passed-in data pid.

Now, when getModel detects its been passed a data PID and it searches for a matching metadata PID, we instead re-navigate to /view/$metadata#data.