Render AnnotationView popover's only on click
Previously, the Bootstrap Popover for each AnnotationView on the page was created when the view's render method was called. This didn't work great when there were hundreds of thousands of AnnotationViews to be created on the page.

Now, we initialize the views we need in MetadataView but only create their Popover on click. On an example dataset, this takes MetadataView.createAnnotationViews' total time from 14,500ms to 85ms.

Closes #1813