Change default behavior of BioPortal tree widget
The previous behavior was that choosing an item from the autocomplete list didn't execute your search and only expanded the tree to the term. Most of us found that to be too many steps. With this change, the widget now executes your search immediately when you click an item in the autocomplete list (configurable). The old behavior is retained (jumping to the right place in the tree) so views that don't clear away the widget (MeasurementTypeView) can select your term and navigate to the term in the tree.

Closes https://github.com/NCEAS/metacatui/issues/1659