Show example layers in Cesium [WIP]
This commit adds example data to the CesiumView, including 3D tiles of
lakes data, markers for LEONetwork permafrost news, plus some publicly
accessible imagery. The CesiumView and associated model will be updated
so that each of these types of layers are configurable, rather than
hard-coded into the CesiumView as they are currently. This commit serves
only to display some example data and is a work-in-progress.

Relates to #1724