Make AnnotationView more backwards compatible
Ref https://github.com/NCEAS/metacatui/issues/1807

Previously, when the new code encountered a DOM structure from an older Metacat installation, there would be some graphical quirks. Short of shipping two separate AnnotationView views for the old structure and the new one, we've decided to disable the view for older Metacat's an make it look good enough.

This code can be removed in the future.