Tweak styling of portal annotation search to address issues
Before this change, styling of the annotation search widget when included in a portal's data search was visually inconsistent with other filters (text, dropdown, etc). This is because the Searchable Select widget has styling that's visually consistent with the new portal editor which has its own look and feel.

With this change, the annotation search widget should look very much at home with other filters.

NB: Some tweaks to the NCBO tree widget were made to improve styling.

Ref #1653
Ref #1842