Move MeasurementTypeView rendering to postRender
Closes https://github.com/NCEAS/metacatui/issues/1896

This moves the rendering of the MeasurementTypeView into the postRender portion of the AttributeView. The postRender method on each AttributeView is called after the view is initially rendered and is called when the user switches to the corresponding tab. The effect is that we render the MeasurementTypeView just in time for the user.