Refactor initial portal semantic search implementation
This tweaks the initial implementation of portal semantic search done in https://github.com/NCEAS/metacatui/pull/1842.

The main work in this commit was to make the semantic search implementation fit in with the rest of the Portals/Filter implementation and close some gaps in the implementation w/r/t the Portal UIBuilder.

In summary:

- Filters that need semantic search now use a SemanticFilterView and rely on a Filter model and its machinery rather than custom state and events.
- Generalized the behavior of having a separate set of labels for values set on the Filter (i.e. show an rdfs:label in the UI when the value is an opaque URI). This adds a new attribute to the Filter model and render-time behavior.
- Semantic portal filters can now be added, edited, and removed whereas they couldn't before.
- Reverted previous changes to the searchableSelect template that were no longer needed because of the above refactor.

Notable changes: This adds an attribute to all Filter models, valueLabels.