Fix URL encoding issues for GET requests for Solr searches
Most of the main menu selections in the UI were afflicted by bad GET
request composition including characters not permitted by the HTTP
standard, which was responsible for the server rejecting the request
with HTTP 400.

These patches resolve the bad request failures by dropping the invalid
injection of backslashes, and substituting special characters with the
URL-encoded representations, primarily square brackets. Lastly, there
was an instance where a Solr query fragment was being composed with
OrcID strings that weren't enclosed in quotes, which is also fixed.

The late string manipulations have been marked as TODO items since a
more permanent, ideal fix is to rework the preceding logic rather than
stripping bad characters back out of a string that was just composed.