Layer List Polish: Add isCategorized attr to LayerItemView
Per discussion [here]( https://github.com/NCEAS/metacatui/issues/2277#issuecomment-1984829655), if a layer list is not categorized, we'll use different colors in the layer icon to indicate layer shown vs hidden. For a categorized layer list, we use the eye icon instead, because a custom icon is already used for the category label.

The different treatments require LayerItemView to be aware of whether it's categorized or not.