Layer List Polish: Widen toolbar panel and layer details view
As a followup to https://github.com/NCEAS/metacatui/pull/2318, we now keep every layer in one line and use ellipsis if the layer name is too long. Unfortunately that is the case for a lot of our layers on the demo site. The fix is to widen the toolbar panel and layer details view so we can accommodate longer layer names.

Then we face a challenge in the main repository maps that use a 3-part view, where the map view doesn't have enough space for the wide panel. Given that it's a setup only existing in main repositories (for now), we'll apply the wide panel only in portal maps.