Fix issues with timing of MetadataView re-render
The metadata view re-renders when a users' logged in status changes. Depending on the timing of this change,
re-rendering may occur in the middle of a previous render. This causes elements of the view to break, especially the
new CanonicalDatasetHandlerView. This commit fixes the issue by adding a check to see if the view is already
rendering before re-rendering.

Issue #2541