Fix paneling issues
Following Robyn's fix related to the Panel issues, the ToobarView.js got replaced with older code. This is now updated with the new version of the code. In addition, there were new issues since the ToolbarView was still using some of the removed HTML classes. In order to fix paneling related issues, I have now replaced all use of the class `draw__tool-panel` with `download-panel`.