Always show new attr button in EMLAttributesView
The button is now a UI element only and is never removed. It is no longer connected to a new blank model in the collection, which was leading to unexpected behaviour.
The button is now always visible, stickign to the bottom of the list container when there are many attributes.

Fixes #799, #1722, #1216